chore: remove the legacy exports from alpha
