+++ plugins/scaffolder-react/src/alpha.ts
@@ -15,4 +15,3 @@
  */
 
 export * from './next';
-export * from './legacy';
