--- plugins/scaffolder-react/src/legacy/extensions/index.tsx
@@ -1,56 +1,0 @@
-/*
- * Copyright 2021 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {
-  LegacyCustomFieldValidator,
-  LegacyFieldExtensionOptions,
-  LegacyFieldExtensionComponentProps,
-} from './types';
-import { Extension, attachComponentData } from '@backstage/core-plugin-api';
-import { FIELD_EXTENSION_KEY } from '../../extensions/keys';
-import { FieldExtensionComponent } from '@backstage/plugin-scaffolder-react';
-
-/**
- * Method for creating field extensions that can be used in the scaffolder
- * frontend form.
- * @alpha
- */
-export function createLegacyScaffolderFieldExtension<
-  TReturnValue = unknown,
-  TInputProps = unknown,
->(
-  options: LegacyFieldExtensionOptions<TReturnValue, TInputProps>,
-): Extension<FieldExtensionComponent<TReturnValue, TInputProps>> {
-  return {
-    expose() {
-      const FieldExtensionDataHolder: any = () => null;
-
-      attachComponentData(
-        FieldExtensionDataHolder,
-        FIELD_EXTENSION_KEY,
-        options,
-      );
-
-      return FieldExtensionDataHolder;
-    },
-  };
-}
-
-export type {
-  LegacyCustomFieldValidator,
-  LegacyFieldExtensionOptions,
-  LegacyFieldExtensionComponentProps,
-};
