+++ plugins/scaffolder/src/deprecated.ts
@@ -212,3 +212,13 @@ export type LayoutTemplate = LayoutTemplateTemp;
  * @deprecated use import from {@link @backstage/plugin-scaffolder-react#LayoutOptions} instead as this has now been moved.
  */
 export type LayoutOptions = LayoutOptionsTemp;
+
+/**
+ * TaskPageProps for constructing a TaskPage
+ * @param loadingText - Optional loading text shown before a task begins executing.
+ * @public
+ * @deprecated - this is a useless type that is no longer used.
+ */
+export type TaskPageProps = {
+  loadingText?: string;
+};
