--- plugins/scaffolder/src/legacy/MultistepJsonForm/FieldOverrides/DescriptionField.tsx
@@ -1,22 +1,0 @@
-/*
- * Copyright 2021 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import React from 'react';
-import { MarkdownContent } from '@backstage/core-components';
-import { FieldProps } from '@rjsf/utils';
-
-export const DescriptionField = ({ description }: FieldProps) =>
-  description && <MarkdownContent content={description} linkTarget="_blank" />;
