--- plugins/scaffolder/src/legacy/TaskPage/TaskErrors.tsx
@@ -1,39 +1,0 @@
-/*
- * Copyright 2022 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { Box } from '@material-ui/core';
-import React, { useEffect, useRef } from 'react';
-import { DismissableBanner } from '@backstage/core-components';
-
-type TaskErrorsProps = {
-  error?: Error;
-};
-
-export const TaskErrors = ({ error }: TaskErrorsProps) => {
-  const id = useRef('');
-
-  useEffect(() => {
-    id.current = String(Math.random());
-  }, [error]);
-  return error ? (
-    <Box>
-      <DismissableBanner
-        id={id.current}
-        variant="warning"
-        message={error.message}
-      />
-    </Box>
-  ) : null;
-};
