+++ plugins/scaffolder/src/next/TemplateEditorPage/DryRunResults/DryRunResultsView.tsx
@@ -27,8 +27,8 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { useDryRun } from '../DryRunContext';
 import { DryRunResultsSplitView } from './DryRunResultsSplitView';
 import { FileBrowser } from '../../../components/FileBrowser';
-import { TaskPageLinks } from '../../../legacy/TaskPage/TaskPageLinks';
-import { TaskStatusStepper } from '../../../legacy/TaskPage/TaskPage';
+import { TaskPageLinks } from './TaskPageLinks';
+import { TaskStatusStepper } from './TaskStatusStepper';
 
 const useStyles = makeStyles({
   root: {
